[WIP] refactor updates out of ObjectPromiseProxy
